spi: xilinx_spi: Reinstate xfer function
When the code was reworked to use the spi memory operation functions
in commit f2dd6599d ("spi: xilinx_spi: Add support for spi memory
operations") the xfer function was removed. However we want to
use the SPI controller to talk to an SD card, so this needs to be
reinstated.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>