hobgoblin: Add proFPGA board support
The proFPGA board has to use a different memory layout to the other
boards we support due to limitations in the Vivido interconnect
generator, namely that it requires memory to be aligned on a natural
boundary. This means that 16Gb of memory cannot be placed at the usual
0x80000000 address.

This means we have to build uboot for the proFPGA board at a different
address. We do not have the same problem for openSBI since later version
are relocatable.

Signed-off-by: David McKay <david.mckay@codasip.com>