net: phy: mscc: Common up 8541/8531 init code
Support for the RGMII_RXID and similar was added in commit <blah> for
the vsc8531 but not for the vsc8541. The code for the two parts is
identical apart from the vsc8531 not supporting [G]MII modes. It is
likely the two parts are actually identical in a different package.
This patch commons up the code into one function as has been done for
much of the rest of the code.

Signed-off-by: David McKay <david.mckay@st.com>