riscv: cheri: Add Zcheri supports to asm.h
Defines macro to reprensent different usage for GPRs and CSR to improve the
code portability for Zcheri extension.
- REG_NUM(r) represents native GPRs with number index. e.g. REG_NUM(0) is
  x0 on non-CHERI and is c0 on CHERI.
- REG(r) represents native GPRs with ABI name. e.g. REG(a0) is a0 on
  non-CHERI and ca0 on CHERI.
- PREG(r) represents native GPRs for storing pointer with ABI name. e.g.
  PREG(sp) is sp on non-CHERI and csp on CHERI.
- PCSR(csr) represents native CSRs for storing pointer with CSR name.
  e.g. PCSR(mepc) is mepc on non-CHERI and mepcc on CHERI.

Add REG, XREG aand PREG load/store macro to improve the code portability
for CHERI extension. The native registers on the core with CHERI
extension is capability register.
- REG is the native register. REG is XREG on non-CHERI and CREG on CHERI.
- XREG is the register for storing native word. XREG is same on both of
  non-CHERI and CHERI.
- PREG is the register for storing native pointer. PREG is XREG on
  non-CHERI and CREG on CHERI.

PC_PTR_L and PTR_L is added. For non-CHERI both are macro of la
pseudo-instruction. For CHERI, PC_PTR_L is llc which derive the offset
pointer with the pcc as the result, the provenance of derived pointer is
same as pcc. It can be used before and after capability being initialized.
PTR_L can be used only after the capability initialization. The capability
pointer provenance is set during capability initialization.