test: Add tests for riscv zcheri extension
ut_zcheri and ut_zcheri_pc_abi are simple test for Zcheri extension.

ut zcheri is a serial of simple test to make sure that the change for RISCV
Zcheri architecture behaves as expected. Currently it is mainly CHERI
exception testing.

ut zcheri_pc_abi is a serial of simple test to test the changes for
Zcheripurecap ABI behaves as expected. Like memory allocator. It also
tests different CHERI capability with correct permission and bounds.