board: codasip: hobgoblin: Support reserved-memory in device tree
For systems using the CMU to manage CHERI tags there are likely to be
reserved memory regions at the top of memory, which conflict with the
address U-Boot will pick for the relocated code etc. So instead use
the LMB (Logical memory blocks) code, initialised from the device
tree, to pick a region of memory for U-Boot's relocated data regions.

This is effectively a copy of the code from
arch/arm/mach-stm32mp/dram_init.c and board/xilinx/common/board.c.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>