configs: codasip-a730-hobgoblin: Enable CONFIG_NET_RANDOM_ETHADDR
Enable the generation of a random MAC address if one isn't set in
the devicetree. Without this U-Boot will crash with a bad pointer
dereference while U-Boot is booting. This is because it calls the
AXI Ethernet driver's stop() function without first calling the
probe().

None of the other drivers appear to have protection against this,
but enabling CONFIG_NET_RANDOM_ETHADDR is pretty much universal.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>