net: xilinx_axi_emac: Fix CHERI register mapping issues
For CHERI we need to map registers using ioremap() to ensure we have
a valid capability. The existing code had been partly converted over to
use ioremap() but in practice this was insufficient as ->iobase (
pointer to the Ethernet registers) was still initialised through a
pointer cast (with no provenance) and ->dmarx (Rx DMA registers) was
initialised in such a way that a) provenance was lost b) would have
had invalid bounds.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>
Fixes: d1087506c923 ("driver: Fix pointer casting")