configs: hobgoblin: Remove setting fdt_high
Because we are passing in a FDT at boot time (rather than having one
embedded in U-Boot or loading it from a file system for example) the
FDT is relocated at boot time into an allocated buffer, which is sized
exactly to fit (see common/board_f.c:reserve_fdt).  In this case we
should not be setting fdt_high to -1, as this prevents
boot_relocate_fdt() from moving the FDT to a suitable location which
will be big enough to allow the FDT to be resized.

Signed-off-by: Stuart Menefy <stuart.menefy@codasip.com>