pci: Write to bridge register as 32 bits
For the Xilinx PCIe controller, it seems to be necessary to write the
bridge control register as a 32 bit write. It seems as if the ECAM
region is somehow not latching the change to the bridge numbers, and is
generating the wrong sort of config cycles.

The Linux kernel has a cryptic comment in that says this register should
be written as a 32 bit number, but why has been lost in the mists of
time. This seems to suggest it has been a problem in the past for some
other controllers.

This patch simply reads the secondary latency timer and then writes the
whole register back as a 32 bit number.

Signed-off-by: David McKay <david.mckay@codasip.com>