dm: core: Change driver data to be uintptr_t
The driver data field is often used to store a pointer, but is defined
as an unsigned long. This will fail for platforms where uintptr_t is a
different length from unsigned long, such as CHERI. This patch also
changes some of the drivers that are casting to ulong instead of
uintptr_t.

Signed-off-by: David McKay <david.mckay@codasip.com>